placeorder refactor

+++ src/main/java/lk/ijse/coir/dao/custom/impl/ItemDAOImpl.java
@@ -20,7 +20,7 @@ public class ItemDAOImpl implements ItemDAO {
             Item item = new Item(
                     resultSet.getString("item_id"),
                     resultSet.getString("item_name"),
-                    resultSet.getDouble("unit_price"),
+                    resultSet.getBigDecimal("unit_price"),
                     resultSet.getInt("qty_on_hand"),
                     resultSet.getString("rawMaterial_id")
 
@@ -78,7 +78,7 @@ public class ItemDAOImpl implements ItemDAO {
     public Item search(String id) throws SQLException, ClassNotFoundException {
         ResultSet rst = SQLUtil.execute("SELECT * FROM item WHERE item_id=?", id);
         while (rst.next()){
-            Item item=new Item(rst.getString(1),rst.getString(2),rst.getDouble(3),rst.getInt(4),rst.getString(5));
+            Item item=new Item(rst.getString(1),rst.getString(2),rst.getBigDecimal(3),rst.getInt(4),rst.getString(5));
             return item;
         }
         return  null;
